+++ include/altacore/ast/import-statement.hpp
@@ -1,6 +1,7 @@
 #ifndef ALTACORE_AST_IMPORT_STATEMENT_HPP
 #define ALTACORE_AST_IMPORT_STATEMENT_HPP
 
+#include <utility>
 #include "statement-node.hpp"
 #include "expression-node.hpp"
 #include "../det/scope.hpp"
@@ -15,15 +16,15 @@ namespace AltaCore {
 
         std::string request;
         bool isAliased = false;
-        std::vector<std::string> imports; // only available on non-alias imports
-        std::string alias;                // only available on alias imports
+        std::vector<std::pair<std::string, std::string>> imports; // only available on non-alias imports
+        std::string alias;                                        // only available on alias imports
 
         std::shared_ptr<DET::Module> $parentModule;
         std::shared_ptr<DET::Module> $importedModule;
         std::shared_ptr<AST::RootNode> $importedAST;
         std::vector<std::shared_ptr<DET::ScopeItem>> $importedItems;
 
-        ImportStatement(std::string request, std::vector<std::string> imports);
+        ImportStatement(std::string request, std::vector<std::pair<std::string, std::string>> imports);
         ImportStatement(std::string request, std::string alias);
 
         virtual void detail(std::shared_ptr<DET::Scope> scope);